On the right track
Added timer to slow down generation so that it is easier to see what is happening.
Currently the algorithm picks the next tile based off of the index of the tiles which I'm pretty sure is incorrect or at the very least inaccurate.

If I can find a way to change this to the actual tiles entropy instead that would be ideal.